backend/requirements.txt
pydantic
fastapi
uvicorn
google-generativeai
python-dotenv
python-multipart
numpy
pandas
scikit-learn
joblib
websockets